requirements.txt
charset-normalizer
colorama
etils
flatbuffers
flax
huggingface-hub==0.27.1
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
mpmath
msgpack
namex
nest-asyncio
networkx
numpy
nvidia-cuda-runtime-cu12
opt_einsum
optax
optree
orbax-checkpoint
packaging
pillow
protobuf
Pygments
PyYAML
regex
requests
rich
safetensors
scikit-learn
scipy
setuptools
simplejson
six
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow_intel
tensorrt
tensorrt_cu12
tensorrt_cu12_bindings
tensorrt_cu12_libs
tensorstore
termcolor
tf_keras
threadpoolctl
tokenizers
toolz
torch
torchaudio
torchvision
tqdm
transformers
typing_extensions
urllib3
Werkzeug
wheel
wrapt
zipp
